refactor!: restructure project to npm workspaces
This commit relocates all backend code to api/ directory, in the form of
an npm workspace. To make the project a bit more maintainable, the db/
directory (which contains the database models that are in use) is also
turned into a workspace, allowing for easier dependency management and
better shareability.