feat(api): implement CQRS pattern
This commit implements the Command Query Responsibility Segregation
(CQRS) pattern in the API.

The pattern separates the read and write operations into separate
services. The write service is responsible for handling commands that
modify the state of the system, while the read service is responsible
for handling queries that retrieve data from the system.

Thus, the new Command and Query services fully replace the previous
utils.js file.

This commit also updates the GET /short/:id and /list endpoints to use
the queue used in Queue-based Load Leveling pattern.